# proposal-builder/docs/OPENAPI.yaml
openapi: 3.0.3
info:
  title: Net & Soft Proposal Builder API
  description: API interna y pública para la creación, administración, publicación y aceptación de propuestas comerciales de Net & Soft Solutions.
  version: 1.0.0
servers:
  - url: https://cotiza.netandsoft.com.ve
    description: Servidor de Producción
  - url: http://localhost:3000
    description: Servidor Local de Desarrollo
paths:
...
  /api/internal/auth/login:
    post:
      summary: Iniciar sesión administrativa con Argon2id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - email
                - password
              properties:
                email:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: Inicio de sesión exitoso. Cookie HttpOnly ns_session emitida.
        '401':
          description: Credenciales incorrectas.
        '423':
          description: Cuenta temporalmente bloqueada por reiterados intentos fallidos.
...
  /api/internal/proposals/{id}/publish:
    post:
      summary: Publicar propuesta creando versión inmutable en JSON (Requiere SALES o ADMIN)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Propuesta publicada exitosamente con nueva versión.